fix: rename file and delete commented code
