+++ src/common/Hooks/useGetCoursesAndAuthors.jsx
@@ -2,7 +2,7 @@ import { useEffect } from 'react';
 import { fetchAllCourses } from '../../store/courses/coureses.actions';
 import { fetchAllAuthors } from '../../store/authors/authors.actions';
 import { useDispatch, useSelector } from 'react-redux';
-import { getAuthors, getCourses } from '../../store/dataFromStore';
+import { getAuthors, getCourses } from '../../store/selectors';
 
 export const useGetCoursesAndAuthors = () => {
 	const dispatch = useDispatch();
